initial scrapping by session update
