+++ package.json
@@ -7,7 +7,7 @@
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "nodemon app.js",
     "populate": "node ./admin/populate.js",
-    "scrape": "node ./admin/scrape.js",
+    "scrape:session": "node ./admin/bysession.js",
     "scrape:speakers": "node ./admin/byspeaker.js",
     "scrape:topics": "node ./admin/topics.js"
   },
